Whenever chaincode changes run the upgrade command instead of an install
